requirements.txt
tensorflow==2.13.0
pillow==8.4.0
numpy>=1.22.0,<1.24.4
